Fixing a runtime.sigpanic() bug (#107)
This package uses cortesi/termlog which uses fatih/color v1.10.0.
I believe color v1.10.0 is causing the panic when testing and
building, but moving to v1.13.0 fixes this. Another option might
be to update cortesi/termlog and then update the version of that
library used here in modd, but this fix is a bit simpler for now
since it is a single commit.